Add db connection message
