+++ deals-front/src/utils.js
@@ -1,0 +1,3 @@
+export const environment = {
+    url: "http://localhost:8090"
+};
